Move py_structure to old folder to start anew
